back-end/pyproject.toml
[tool.poetry]
name = "quickshare"
version = "0.1.0"
description = "A privacy centered p2p photo sharing platform with an added storage functionality"
authors = ["Jason Cameron <git@jasoncameron.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
aiohttp = "^3.9.5"
aiofile = "^3.8.8"
gcloud-aio-storage = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
